moved components to sampling folder
