+++ node/src/components/sampling/JobSampling.tsx
@@ -1,12 +1,12 @@
 import { InfoOutline } from "@mui/icons-material";
 import { Accordion, Button, AccordionDetails } from "@mui/material";
 import React, { useState } from "react";
-import CustomTooltip from "./CustomTooltip";
-import AdaptExtedSamplingDocument from "./documents/AdaptExtendSamplingDocument";
-import JobSelector from "./JobSelector";
+import CustomTooltip from "../CustomTooltip";
+import AdaptExtedSamplingDocument from "../documents/AdaptExtendSamplingDocument";
+import JobSelector from "../JobSelector";
 import PlusButton from "./PlusButton";
 import { Sampling } from "./Sampling";
-import { Function } from "../osparc-api-ts-client";
+import { Function } from "../../osparc-api-ts-client";
 
 interface JobSamplingProps {
   loading: boolean;
