+++ node/src/components/sampling/RunSamplingButton.tsx
@@ -1,8 +1,8 @@
 import { Button, CircularProgress } from '@mui/material';
-import { useMMUXContext } from '../context/MMUXContext';
-import { useServiceContext } from '../context/ServiceContext';
+import { useMMUXContext } from '../../context/MMUXContext';
+import { useServiceContext } from '../../context/ServiceContext';
 import { toast } from 'react-toastify';
-import CustomTooltip from './CustomTooltip';
+import CustomTooltip from '../CustomTooltip';
 
 type RunSamplingButtonProps = {
   handleRunSampling: () => void;
